Move alCheck macrofunction into its separate files
