+++ rwengine/src/audio/SoundManager.cpp
@@ -1,16 +1,10 @@
 #include <audio/SoundManager.hpp>
 #include <audio/MADStream.hpp>
 
+#include "audio/alCheck.hpp"
+
 #include <array>
 #include <iostream>
-
-void checkALerror(const std::string& file, unsigned int line);
-
-#if RW_DEBUG
-	#define alCheck(stmt) do { stmt; checkALerror(__FILE__, __LINE__); } while(0)
-#else
-	#define alCheck(stmt) stmt
-#endif
 
 void SoundManager::SoundSource::loadFromFile(const std::string& filename)
 {
@@ -140,29 +134,5 @@ void SoundManager::pause(bool p)
 		} else {
 			playSound(backgroundNoise);
 		}
-	}
-}
-
-void checkALerror(const std::string& file, unsigned int line)
-{
-	ALenum err = alGetError();
-	if (err != AL_NO_ERROR) {
-		std::cerr << "OpenAL error at " << file << ":" << line << ": ";
-
-		switch (err) {
-		case AL_INVALID_NAME:
-			std::cerr << "Invalid name!";
-			break;
-		case AL_INVALID_VALUE:
-			std::cerr << "Invalid value!";
-			break;
-		case AL_INVALID_OPERATION:
-			std::cerr << "Invalid operation!";
-			break;
-		default:
-			std::cerr << err;
-		}
-
-		std::cerr << std::endl;
 	}
 }
